+++ modules/mola_input_euroc_dataset/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package mola_input_euroc_dataset
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.2.2 (2023-09-08)
+------------------
+* Correct references to license
+* Contributors: Jose Luis Blanco-Claraco
+
 0.2.1 (2023-09-02)
 ------------------